- Implemented CustomEndpointResponse [1/4]
Summary:
Following this [Design Doc](https://docs.google.com/document/d/1_19vhaezTArpba783Qaxv_h8QjJubBbWX06yJllSdpQ/edit)

Implementing the return type of SendEvent and SendEventAsync.
Class highlighted in red below
{F778708950}

This class is used as a field in EventResponse to pass to the user that this is the result of the endpoint request.
The key is string which is the URL and the value is the response result containing the message and response code.

Reviewed By: ztlbells

Differential Revision: D40162781

fbshipit-source-id: ed7824943929b03138f6d94e76234e3732e06cb4