Added unit tests to EventRequest executeSync and executeAsync functions [3/4]
Summary:
This is the last planned diff for the Business SDK java implementation. Writing a diff for the EventRequest changes. The components being tested in this diff are highlighted in red in the below diagram.
{F780372421}

Unit Tests Coverage:

- Sending a synchronous event to both the Custom Endpoint and CAPI
- Sending a synchronous event to both the Custom Endpoint only
- Sending an synchronous event to Custom Endpoint only

Next steps: Starting work on the PHP Implementation and preparing for the Java and PHP release.

Reviewed By: ztlbells

Differential Revision: D40295143

fbshipit-source-id: c7ebdf8dd8cfa893e660e66bed9459ca0fcb0b0f